+++ packages/orq-rc/src/models/operations/updatedatapoint.ts
@@ -152,13 +152,13 @@ export type UpdateDatapointMessagesDatasetsContent =
   | Array<UpdateDatapoint2TextContentPart | UpdateDatapoint2RefusalContentPart>;
 
 /**
- * The role of the messages author, in this case `assistant` or `exception`.
+ * The role of the messages author, in this case `assistant`.
  */
 export const UpdateDatapointMessagesDatasetsRequestRequestBodyRole = {
   Assistant: "assistant",
 } as const;
 /**
- * The role of the messages author, in this case `assistant` or `exception`.
+ * The role of the messages author, in this case `assistant`.
  */
 export type UpdateDatapointMessagesDatasetsRequestRequestBodyRole = ClosedEnum<
   typeof UpdateDatapointMessagesDatasetsRequestRequestBodyRole
@@ -226,7 +226,7 @@ export type UpdateDatapointMessagesAssistantMessage = {
    */
   refusal?: string | null | undefined;
   /**
-   * The role of the messages author, in this case `assistant` or `exception`.
+   * The role of the messages author, in this case `assistant`.
    */
   role: UpdateDatapointMessagesDatasetsRequestRequestBodyRole;
   /**
@@ -668,14 +668,14 @@ export type UpdateDatapointMessagesDatasetsResponse200Content =
   >;
 
 /**
- * The role of the messages author, in this case `assistant` or `exception`.
+ * The role of the messages author, in this case `assistant`.
  */
 export const UpdateDatapointMessagesDatasetsResponse200ApplicationJSONResponseBodyRole =
   {
     Assistant: "assistant",
   } as const;
 /**
- * The role of the messages author, in this case `assistant` or `exception`.
+ * The role of the messages author, in this case `assistant`.
  */
 export type UpdateDatapointMessagesDatasetsResponse200ApplicationJSONResponseBodyRole =
   ClosedEnum<
@@ -745,7 +745,7 @@ export type UpdateDatapointMessagesDatasetsAssistantMessage = {
    */
   refusal?: string | null | undefined;
   /**
-   * The role of the messages author, in this case `assistant` or `exception`.
+   * The role of the messages author, in this case `assistant`.
    */
   role:
     UpdateDatapointMessagesDatasetsResponse200ApplicationJSONResponseBodyRole;
@@ -6252,7 +6252,7 @@ export const UpdateDatapointEvaluations3$inboundSchema: z.ZodType<
     .default("orq"),
   reviewed_by_id: z.string(),
   reviewed_at: z.string().datetime({ offset: true }).default(
-    "2025-10-16T10:40:29.535Z",
+    "2025-10-16T14:53:50.205Z",
   ).transform(v => new Date(v)),
   type: UpdateDatapointEvaluationsDatasetsResponseType$inboundSchema,
   values: z.array(z.string()),
@@ -6290,7 +6290,7 @@ export const UpdateDatapointEvaluations3$outboundSchema: z.ZodType<
   source: UpdateDatapointEvaluationsDatasetsResponseSource$outboundSchema
     .default("orq"),
   reviewedById: z.string(),
-  reviewedAt: z.date().default(() => new Date("2025-10-16T10:40:29.535Z"))
+  reviewedAt: z.date().default(() => new Date("2025-10-16T14:53:50.205Z"))
     .transform(v => v.toISOString()),
   type: UpdateDatapointEvaluationsDatasetsResponseType$outboundSchema,
   values: z.array(z.string()),
@@ -6419,7 +6419,7 @@ export const UpdateDatapointEvaluations2$inboundSchema: z.ZodType<
   source: UpdateDatapointEvaluationsDatasetsSource$inboundSchema.default("orq"),
   reviewed_by_id: z.string(),
   reviewed_at: z.string().datetime({ offset: true }).default(
-    "2025-10-16T10:40:29.535Z",
+    "2025-10-16T14:53:50.205Z",
   ).transform(v => new Date(v)),
   type: UpdateDatapointEvaluationsDatasetsType$inboundSchema,
   value: z.number(),
@@ -6458,7 +6458,7 @@ export const UpdateDatapointEvaluations2$outboundSchema: z.ZodType<
     "orq",
   ),
   reviewedById: z.string(),
-  reviewedAt: z.date().default(() => new Date("2025-10-16T10:40:29.535Z"))
+  reviewedAt: z.date().default(() => new Date("2025-10-16T14:53:50.205Z"))
     .transform(v => v.toISOString()),
   type: UpdateDatapointEvaluationsDatasetsType$outboundSchema,
   value: z.number(),
@@ -6581,7 +6581,7 @@ export const UpdateDatapointEvaluations1$inboundSchema: z.ZodType<
   source: UpdateDatapointEvaluationsSource$inboundSchema.default("orq"),
   reviewed_by_id: z.string(),
   reviewed_at: z.string().datetime({ offset: true }).default(
-    "2025-10-16T10:40:29.535Z",
+    "2025-10-16T14:53:50.205Z",
   ).transform(v => new Date(v)),
   type: UpdateDatapointEvaluationsType$inboundSchema,
   value: z.string(),
@@ -6617,7 +6617,7 @@ export const UpdateDatapointEvaluations1$outboundSchema: z.ZodType<
   humanReviewId: z.string(),
   source: UpdateDatapointEvaluationsSource$outboundSchema.default("orq"),
   reviewedById: z.string(),
-  reviewedAt: z.date().default(() => new Date("2025-10-16T10:40:29.535Z"))
+  reviewedAt: z.date().default(() => new Date("2025-10-16T14:53:50.205Z"))
     .transform(v => v.toISOString()),
   type: UpdateDatapointEvaluationsType$outboundSchema,
   value: z.string(),
@@ -6761,7 +6761,7 @@ export const UpdateDatapointResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-16T10:40:20.235Z",
+    "2025-10-16T14:53:40.726Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -6840,7 +6840,7 @@ export const UpdateDatapointResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-16T10:40:20.235Z"))
+  updated: z.date().default(() => new Date("2025-10-16T14:53:40.726Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {